Update client-side local storage utils
